Added setting bg image in current weather item plu developed layout and color theme

+++ src/main/java/com/weather/controller/MainWindowController.java
@@ -61,7 +61,7 @@ public class MainWindowController {
 
         //weather for the next four days:
         for (int i = 1; i <= 4 ; i++) {
-            temps = String.valueOf(weathers.get(i).getDayTemperature()) + "°C / " +  String.valueOf(weathers.get(i).getNightTemperature()) + "°C";
+            temps = String.valueOf((int)weathers.get(i).getDayTemperature()) + "°C / " +  String.valueOf((int)weathers.get(i).getNightTemperature()) + "°C";
             description = weathers.get(i).getDescription();
             date = String.valueOf(i);
             Node node = viewFactory.createWeatherItem(temps, description, date);
